requirements.txt
fastapi==0.109.1
uvicorn==0.27.0.post1
google-cloud-firestore==2.14.0
pydantic-settings==2.1.0

# test
mock-firestore==0.11.0